TYP: Don't narrow BlockContent return type in the subclasses
If the return type of class methods of the Render* classes is narrower
(explicity by annotating with a narrower type OR implicitly by type of
the returned value) than the respective methods in RenderBase; the
methods in the classes subclass the Render* classes must also use the
narrower types. We don't want that.